Merged PR 1387026: add symbol files to our nuget package.  pack.sh already is using the
add symbol files to our nuget package.  pack.sh already is using the --symbols switch so that it will produce a public .nupkg as well as a .symbols.nupkg file, and we are already only pushing the .symbol.nupkg file to the internal VSO feed.  unfortunately neither currently contains symbol files.  This change fixes that.

We can continue to use the base .nupkg file when uploading to nuget.org, and the .symbols.nupkg for windows shell.
Still not sure what this means for windows moving back to downloading the package from nuget.org instead of our VSO nuget feed.

Related work items: #2193849