Merged PR 1375249: Fixing native library crash issue with spring boot
Fixing native library crash issue with spring boot.
ClassLoader.getSystemClassLoader().getResource(".") was crashing in java springboot application. Also according to the description in the comment the pma.dll should be in the same folder as the executable jar. while ClassLoader.getSystemClassLoader().getResource(".").getPath() was getting the $(jarPath)\classes.

So making the code changes is solving the issue on spring boot. We need to test with pma.dll though.

Related work items: #2180090