Merged PR 1394089: add exception handling to CPP quick starts, and add missing solution file to
Jakoby Etchells reported that the windows samples compiled, but crashed on execution.
After a few attempts, we found that there was not microphone\device plugged in, and the quickstarts do not have exception handling to report a usefull error to the user.

This PR adds rudimentery exception logging, so the user can see "SPXERR_MIC_NOT_AVAILABLE" or "SPXERR_FILE_OPEN_FAILED", and figure out how to fix the problem.

Related work items: #2195621