Merged PR 1457138: Improving ref docs for DialogServiceConnector
This is part of an effort to improve the quality of the reference docs that the SDK has. This change has the intention to explore on how to make them better and establish that as a guide to the rest of Carbon.

Concretely, this change the documentation on FromConfig. The result of the refactor yielded this refdoc entry:

![ss.png](https://msasg.visualstudio.com/e71f1362-9c7d-488b-99c7-3376db8d3302/_apis/git/repositories/c01407a1-8dbe-4bbf-8f73-668d96a579c2/pullRequests/1457138/attachments/ss.png)

One missing thing for this is to check how this would look once pulled into Azure's ref docs.

Related work items: #2285900