Merged PR 1472472: Fix the symbolic linking in CMake to do the right thing if the symbolic link cannot be created
If a user is not running the cmake with admin, the symbolic links fail to be created silently. There are 2 ways to fix this, way one is to fail cmake fatally, the other is to copy the file rather than fail. I have, provisionally selected option 2 and tested it for feasibility. It works for normal users, and so I have implemented it with if fail to link copy instead

Related work items: #2303800