Merged PR 1479299: Calling convention for Android C API does not match with C# dllimport (cdecl vs stdcall)
Based on investigation done by Xamarin team for the issue: https://github.com/xamarin/xamarin-android/issues/3989

DllImports in the Microsoft.CognitiveServices.Speech.csharp.dll are marked as stdcall, however the native code from libMicrosoft.CognitiveServices.Speech.core.so is cdecl (default) for Android. This inconsistency in the call convention causes the stack misalignment in Android x86 platform and e.g. Android x86 emulator crashes when used. On arm architecture, the call convention is standardized by the ABI and you can’t change it. The convention that is passed in C# in the attribute will be ignored by the runtime on arm.

The change is to enable stdcall in Android x86 platform and verified that this fixes the issue with Android x86 emulator.

Related work items: #2299173