Merged PR 1504103: Fix for kws when usp connection error happens (GitHub issue 506)
When KWS is active and service connection timeouts(after 10 minutes), KWS functionality stops working so that speech is not recognized anymore. The fix is add a check in error handling when we the session is in Keyword mode, Canceled event is not sent, but we ignore the error and ensure reco adapter is reseted when next time speech recognition starts via keyword activation.

https://github.com/Azure-Samples/cognitive-services-speech-sdk/issues/506

Related work items: #2350335