Merged PR 1537225: Support C# on Linux ARM32 and ARM64
Updated the nuget package and related configuration so that .NET Core C# can be used on Linux ARM32 and ARM64. Verified that quickstarts and samples work on Raspberry Pi 4 with Ubuntu 18.04 for ARM32 (dotnet 2.1 and 3.1) and ARM64 (dotnet 3.1).

Docker images were created for automated testing, but the current versions are not used in CI due to incompatibilities:
1. oobedevdnc21_ubuntu1804_arm32 (with dotnet 2.1, also tried 3.1 with the same results) produces various qemu errors already on 'dotnet restore' despite the latest qemu version
   - "qemu: uncaught target signal 4 (Illegal instruction)"
   - "qemu: uncaught target signal 6 (Aborted)"
   - "qemu: uncaught target signal 11 (Segmentation fault)"
2. oobedevdnc31_ubuntu1804_arm64 (with dotnet 3.1 for proper Linux ARM64 support) works locally but
   - occasionally spews errors "shared memfd open() failed: Function not implemented" apparently from pulseaudio
   - and needs netcoreapp3.0 or newer as TargetFramework while CI and samples currently use netcoreapp2.0.
   When tried on HW, it is actually possible to have netcoreapp2.0 by first installing .NET Core 2 runtime and then .NET Core 3 SDK on top of that, but in Docker this combination just makes dotnet get stuck.

Some general changes were made in NuGetLinuxTest to prepare for ARM testing when possible in the future. Also updated oobedevdnc21_centos8_x64 image so that it now works in CI (previously disabled).

Related work items: #2074653, #2074655