Merged PR 1613727: Better error & exception handling for Java JNI layer
For most of the JNI calls, need to check if there is pending exceptions. This PR does that, currently clears the exception if there is any. This PR also adds more error handling to JNI layer and also wraps UTF string get and release methods to be more maintainable.

Related work items: #2159909