Merged PR 1680592: Add support for setting authorization token on the ConversationTranslator
For the conversation translator scenario, all participants in the conversation will use the speech subscription for the host. They don't need to know the subscription information when joining the conversation since the join call will return an authorization token to use.

However, in the case where the host created the room using an authorization that expires while the conversation is still active, this will prevent everyone in the room from being to do speech recognition. This updates the Carbon SDK to properly parse the new "an updated authorization token" message that is broadcast to all participants in the room. It also add support for the authorization token set by the host to be sent up to the conversation service to be sent to all participants.

Also includes the following bug fixes and/or changes:
* The utterance IDs were being inconsistently being converted to upper cases which was causing issues in client code
* Added a proprety/getter for your current participant ID since this was impossible to determine until you got a transcription back for you (or you sent an IM and got back the translations)
* Both authorization token and subscription key were being set on requests. Instead now only one will be set

Related work items: #2841626