Merged PR 1695968: Make SendMessageAsync finished sending the message after users finish wait for it.
Added a promise to each TransportPacket. The promise is signaling when azure-c-shared-lib calls the callback, aka WebSocket::HandleWebSocketFrameSent, of the message being sent over socket. The future of the promise is passed back to the audio stream session, so that we can wait for the result there.

Added a "send_event_without_audio" test section under test case of "Dictation Corrections" in speech_recognizer_tests.cpp.

Related work items: #2231116