Merged PR 1664982: Add a null check to the recently-added reconnect-upon-error logic
This follows up on !1562963 by adding a null check before the use of `m_audioBuffer` at the start of the function.

`m_audioBuffer` **does** have more naked derefs inside of this same function, but:

1. This is the only new addition
2. This is the one being hit
3. The error condition responsible for the reported crash *does not* go down any of the other existing code paths that deref the buffer
4. In general, the call within the reconnect loop is a different code path

What makes this fundamentally confusing is that `::Error` is called for too many different reasons. This could really benefit from a future refactor/redesign, but that's beyond the scope of fixing the initial crash.

Related work items: #2806491