Merged PR 1722972: Add Connection::FromDialogServiceConnector
`DialogServiceConnector` is currently inconsistent in that it has `ConnectAsync` and `DisconnectAsync` functionality built in to the top level of the object. Everything else manages connection state via `Connection.From*()`. That on its own is an area for improvement but not necessarily a huge deal; the bigger problem is that we *don't* have the ability to get a `Connection` object from the `DialogServiceConnector` and the root object itself *does not* have the `Connected` and `Disconnected` events that allow an application to respond to service connection status changing.

Having this exposed the right way is a prereq to fixing #2876682 since we still need to expose connection state to the developer in some way -- ideally just not as an error.

Related work items: #2876682