Merged PR 1741320: Allow end users to set any HTTP header key/value pairs.
There will be two ways to do it.
1)  config->SetServiceProperty("Authorization", "FromSpeechConfig", ServicePropertyChannel::HttpHeader);
2) recognizer->Properties.SetProperty("HttpHeader#Authorization", "fromRecognizer");

Internally, Carbon prefixes the ServicePropertyChannel enum to the name ( such as "Authorization"). So, in the property bag there is "HttpHeader#Authorization" as "key" and its value is "FromSpeechConfig" for case 1.

This PR also hooks up the correct way of accessing properties in CSpxUspRecoEngineAdapter class, so that from CSpxUspRecoEngineAdapter, we can access the properties live in one of its sites in the filter chain.

Also, move the implementation of property bag into cpp, adding property_bag_impl.cpp.

Related work items: #2899459