Merged PR 1770670: Update the flag for MSC bug workaround
There is a bug in the Microsoft C++ compiler that prevents from creating a packaged_task from a move only lambda: https://github.com/microsoft/STL/issues/321. This makes the compilation of CSpxUspTtsEngineAdapter::UspSendMessage() fail:

`
7>usp_tts_engine_adapter.cpp
1>MSBuild auto-detection: using msbuild version '16.7.0.37604' from 'C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Current\Bin'.
7>C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\VC\Tools\MSVC\14.27.29110\include\functional(879,58): error C2280: 'Microsoft::CognitiveServices::Speech::Impl::CSpxUspTtsEngineAdapter::UspSendMessage::<lambda_2a9ddf0ef56cc1937697c8ec575a9c4c>::<lambda_2a9ddf0ef56cc1937697c8ec575a9c4c>(const Microsoft::CognitiveServices::Speech::Impl::CSpxUspTtsEngineAdapter::UspSendMessage::<lambda_2a9ddf0ef56cc1937697c8ec575a9c4c> &)': attempting to reference a deleted function
`

The code in question already had a workaround but it was under "#if _MSC_VER <= 1926" which means the same error pops up after VS2019 version 16.7 update.

Instead of increasing the version number in the condition above every time VS2019 is updated, it's better to enable the workaround by default for now.

Related work items: #2937055