Merged PR 1770724: Sanity now: ensure unreliable test waits for stopContinuousRecognitionAsync to finish
The error in question:

`java.lang.IllegalStateException: Cannot dispose a recognizer while async recognition is running. Await async recognitions to avoid unexpected disposals.`

The cause of this is pretty clear--`r.close()` is invoked immediately after the call to `stopContinuousRecognitionAsync()`, which is wrong. Simply waiting (with a timeout) for the future should resolve this.

Related work items: #2934773