Merged PR 1777863: ContinuousValidSkipAudioRecognition is failing after recent service changes.
seems this test was cutting the audio right in the middle of the word THE.
after a service change, the service did a better job of rejecting the half work in the start of the reco.
But could reasonably change back to the previous behavior give we sent half a word.

To make the test more stable, I am backing the trimmed audio to be clearly in-between the words "What's" and "the".  this should make the service deterministic and if their is a future failure, could be a real bug in the service or carbon.

Related work items: #2943380