Merged PR 1783149: Increase Python test timeout for session stopped event
In Python tests, _wait_for_event(callbacks, 'session_stopped') occasionally fails e.g. in
https://msasg.visualstudio.com/Skyman/_build/results?buildId=14622361&view=results

The current timeout setting for that is
```
# the timeout to wait for session stopped event after recognition is finished
_TIMEOUT_IN_SECONDS = 10.
```
It's a known issue that the session stopped event can take its time, and there's task 2943221 to improve the handling of this on the Carbon side when the service doesn't end the turn fast enough. To reduce the probability of this timeout occurring during CI builds for now, I increased the value used in Python tests to 30 seconds.

Related work items: #2948485