Merged PR 1788976: Migrate c++ tests to a new .trx reporter for better CI output
## Background/approach

Catch2 uses a concept of a "reporter" for what and how it emits results during and/or after a test run. Making new ones of these is explicitly [not documented on the project](https://github.com/catchorg/Catch2/blob/master/docs/reporters.md#top) with "you can figure it out" as the order of the day. Reporters derive from one of two bases:

* `CumulativeReporterBase`, as used by `JUnitReporter`, which is designed to emit a DOM-style monolith "after the fact"
* `StreamingReporterBase`, as used by Catch2's `XmlReporter`, which is designed to do a progressive emission as the tests run

This change introduces a `VstestReporter` that emits an XML file that mimics/follows the .trx file format that we get from our C# runs.

The concepts in Vstest and the way this reporter is written seem much better suited to a cumulative reporter model, but it's implemented (for now) atop the streaming reporter model as that gives the most flexibility in what information we store and how we store it--including the flexibility to deal with things like the "wind-up" of section starts not necessarily aligning with the "unwind" of section ends in cases like exception-related failures.

At a high level, Catch2 "winds up" with starts to an arbitrary number of embedded test sections (including one at the top-level of a test case) and then "winds down" with matching section ends. It repeats the runs of parent sections for every leaf/child section,  meaning that the fundamental unit we can represent as a test case is a full wind/unwind; we don't need to worry about a branching topology. That's hard to follow, so a more visual example:

```
TEST CASE START
  ASSERTION 1
  SECTION A
    ASSERTION 2
    SECTION A1
      ASSERTION 3
    SECTION A2
      ASSERTION 4
    ASSERTION 5
  SECTION B
    ASSERTION 6
  ASSERTION 7
TEST CASE END
```

This ends up with parses that look like:
```
Start -> Assertion 1 -> Section A -> Assertion 2 -> Section A1 -> Assertion 3 -> Assertion 5 -> Assertion 7 -> End
Start -> Assertion 1 -> Section A -> Assertion 2 -> Section A2 -> Assertion 4 -> Assertion 5 -> Assertion 7 -> End
Start -> Assertion 1 -> Section B -> Assertion 6 -> Assertion 7 -> End
```

Which is three effective test cases.

In terms of grouping, this Catch2 structure would support arbitrary levels of nesting, but .trx (despite its schema) seems to really only support a single layer of nesting--you can have a top-level "DataDrivenTest" with a number of child "DataDrivenTestRow" cases, but nothing beyond that. To fit this, all parses are collapsed as children of a parent test for the whole case. For the above, that means we have:

```
Test Case [Overall pass/fail + aggregate stats]
-> Section A / Section A1 [pass/fail]
-> Section A / Section A2 [pass/fail]
-> Section B [pass/fail]
```

## Future work / to-do

* This is currently accomplished by generating the single header out of a fork of the official Catch2 project: http://www.github.com/trrwilson/Catch2
* There'...