Merged PR 1812975: Add some PushStream tests
We have a customer who has expressed interest in using the functionality of writing an empty buffer to a PullStream and then re-using the recognizer and adding more data later to continue recognition.

While it works, I couldn't find a test for this behavior, so adding some to ensure we don't break them later.

Related work items: #2979857