Merged PR 1830837: Remove defunct server latency test from pipeline
This test runs only on scheduled builds and consistently failed for as long as we have records. Yes, that means every scheduled build we've had has failed for this (assuming it got this far and didn't fail in a previous step like Run Tests) and we didn't notice it until now.

I filed this item to track consideration of reintroducing the tests:
#2993523

Related work items: #2993628