Merged PR 1826966: Back up catch result files in the event that they're malformed
Some circumstances can cause the entire test process to crash catastrophically *outside* of the fatal signal unwind that's built in -- e.g. this concurrency issue we identified:
#2993922

We should address those conditions individually, but we ultimately should have a safe way of ensuring that -- no matter what -- we still have the test output available for diagnosis in some form when the "nice stuff" fails.

This change simply adds a new artifact to the build with .zip files for each of the configuration's .trx files. That'll included a malformed .trx in the event that it's malformed, meaning we may still get some output on a test case if it had a chance to emit before the crash.

![image.png](https://msasg.visualstudio.com/e71f1362-9c7d-488b-99c7-3376db8d3302/_apis/git/repositories/c01407a1-8dbe-4bbf-8f73-668d96a579c2/pullRequests/1826966/attachments/image.png)

Related work items: #2971073