Merged PR 1836655: Fix variable expansion and directory path issues in Proxy test runs.
there are two issues causing the proxy tests to fail.
the first is that the "${SPEECHSDK_SPEECH_KEY}" variable was being expanded on the VS Build host, instead of inside the running docker container.

The second was that the current working directory was wrong, as well as an incorrect path to the test input wave files used by the test.

Related work items: #2998109