Merged PR 1836619: Fix mangled test output from garbled CDATA processing in uspclientconsole, others
The previous change for this addressed one of *two* problems with the "nice but unnecessary" move/simplification of `tail` use inside of `test-harness.sh`. The second one involved the way that error contents get tokenized inside of the `printf` that's forming the string, which caused spurious tokens to barf all over subsequent test cases. I'm not fully sure why this is (something intrinsic to bash) but fortunately also don't need to fully understand it to fix it.

Putting the error output back into a double-quote enclosed temporary variable (rather than inline backtick-enclosed execution) fixes it.

**Additionally**: I saw no reason for uspclientconsole tests to be filtered to a single test on PR/dev builds. It's a very fast suite and it does occasionally catch legitimate issues. So I removed the filter, which will make the uspclientconsole suite run consistently across all builds.

Example of a test results page with a forced repro of the issue:

https://msasg.visualstudio.com/DefaultCollection/Skyman/_build/results?buildId=15234505&view=ms.vss-test-web.build-test-results-tab

Example of a test results page with a forced repro and the fix applied:

https://msasg.visualstudio.com/DefaultCollection/Skyman/_build/results?buildId=15234626&view=ms.vss-test-web.build-test-results-tab

Related work items: #2995475