Merged PR 1842840: Still issues with passing the speech key and region to the mac os proxy test bash script.
Despite this working locally, an in cloud builds for linuxproxytests, this is still failing on MacOS.
I've tried everything under the sun, and it still failed, despite using Echo to log the $SPEECHSDK_SPEECH_KEY just before and just after calling osx-proxy-test.sh.

As best I can tell, there is something about osx-proxy-test.sh or the fact it is running on OSx rather than linux/windows, that is running the osx-proxy-test.sh script or the call inside it to run CarbonX as a seporate process that is not getting passed the speech key and region variable.

On a wim, I used export to flag these two variables to be passed to "child processes" and that resolved the issue, but I'm not clear why this is needed here, and not in the 100 other place we call a bash script without calling "export" on all the shell variable that the bash scripts use internally.

Related work items: #2998109