Merged PR 1845745: Add some concurrency protection for most Catch2 calls when used via SPXTEST_ macros
## Problem
Catch2 isn't thread-safe. Known design limitation. You can very trivially reproduce the consequences of this with a simple test like the following:

```cpp
TEST_CASE("multithreading")
{
    auto num_threads = 2; // std::thread::hardware_concurrency();
    std::vector<std::thread> pool;
    for (unsigned int i = 0; i < num_threads; i++)
    {
        pool.push_back(std::thread([]
            {
                for (int i = 0; i < 10000; i++)
                {
                    REQUIRE(true);
                }
            }));
    }
    for (auto& t : pool)
    {
        t.join();
    }
}
```

This will crash and burn in a variety of spectacular ways with even 2 threads (waay before 10000 iterations in my experience).

So what? Well, a lot of our functionality is event-driven: we get callbacks for things like recognitions that we perform checks on--these can happen at the same time as other events or other activity not related to events. That opens us up to a bunch of potential cases where we could encounter the same concurrency issue.

At best, it means the test crashes. At worst, heap corruption can do lots of crazy things--I *suspect* (but haven't confirmed) that it can mangle our test output and/or remove redirect of stderr from output. It's generally just very bad™ and we don't want it.

## Approach
Two main options:

1. Modify Catch2 itself. The origin project is C++98-based, which makes things ickier than we'd like We could add in conditional compilation flags that introduce a C++11 dependency, but that's not necessarily going to fly either. It's also trickier to get right.
2. Modify the `SPXTEST_*` helpers that we wrap Catch2's macros with. We want to ensure we use these *anyway*, so there's an ancillary benefit in "encouraging" us to adapt/switch sooner.

This goes down the road of (2). Which doesn't preclude (1) later.

In particular:

* The existing SPXTEST_ wrappers for one-line macros (REQUIRE, CHECK, etc.) are all wrapped with the use of a static mutex
* Wrappers for code block macros (TEST_CASE, SECTION, GIVEN, etc.) are not -- I believe this can be done but it can be a future improvement

## Validation
Using the code above as a basis, adding a static mutex (outside the method) and modifying the for loop like so:
```cpp
for (int i = 0; i < 10000; i++)
{
  std::lock_guard<std::mutex> lockyLock{ g_concurrencyMutex };
  REQUIRE(true);
}
```

...Makes everything happy. All the threads (even 16 if I uncomment the `hardware_concurrency` line) work as expected. So in principle, just sharing a mutex between these should make a *big* difference in concurrent stability.

Related work items: #2993922