Merged PR 1855799: Switch CiCd from "Alternate" to updated "Primary" ESRP Service connection.
CiCd uses a client cert to authenticate to the windows ESRP code signing lab.
Once a year we have to renew our client cert to continue to make signing requests during out builds.
This cert is stored in a one of two secure ADO "Service Connection" in the build pipeline.
We then rotate through the two connections after validating in a manual/PR build.

I've already validated this Cert and Service connection in the SPX build pipeline, and in this manual build in carbon. https://msasg.visualstudio.com/Skyman/_build/results?buildId=15480009&view=results

This change rotated carbon CiCd over to the "Primary" service connection.
The "alternate" in peoples over PRs and previous release branches will continue to work until the cert in the "alternate" connection expires in two weeks.  After that you will need to pull or cherry pick this change to have code signing function on any branch.

Side note: this is the commit to more the SPX builds to use the Primary ESRP service connection too.

92420468 ("ESRP Certs are expiring, this rotates to the Service connection with the renewed certs.", 2020-09-24)

Related work items: #3014593