Merged PR 1859807: Move to service controlled offset caculations
This is work that was in flight from before the CI/CD freeze that the CTS team needs.
(And frankly, was difficult to test given noise levels)

This moves the calculation for offsets from the audio stream session down to the reco adapters, and also introduces a method to indicate that audio has been processed and will not be needed for replay (instead of depending on inspection of final results to determine audio was no longer needed for replay.)

In the event the service is using header offset / acknowledgement control (CTS for now, ETA by EOY for other endpoints) use those headers instead of JSON inspection.

Related work items: #2826729, #2826733