Merged PR 1864911: Give nuget job an extra 10 minutes to finish.
The nuget job timed out in tonight nightly build.

The current limit was set to 30 minutes.
actual runtime on the last half dozen successful builds was 21-25 minutes.
Give that there are multiple network operation, including a request to the ESRP signing lab, it probobly reasonable to give this job an extra 10 minute to wait for network delays and ESRP request backups.

https://msasg.visualstudio.com/Skyman/_build/results?buildId=15632095&view=logs&j=67ea8d94-b5a7-504b-2d4b-07baddfad3f1&s=d654deb9-056d-50a2-1717-90c08683d50a

Related work items: #3024937