Merged PR 1870225: Force brew to not cleanup
Force brew to not cleanup.

We're 31.75 days into the life the MacOs images used in the lab.
On day 30, brew automatically uninstalls "old" things.

One of which was an ancient OpenSSL version. That it looks like Python wanted.

Long (like next couple of weeks) we should figure out how to remove that dependency.

Related work items: #3030570