Merged PR 1867708: Re-enable reconnect support
Reconnect support was broken a while back if the service had politely closed the connection.

This fixes it and adds a test case using CTS.

What will happen now is that for a either a disconnect or an error an error payload object will be sent from the reco engine adapter to the audio stream session either via the Error() or the AdapterDisconnected() methods. (For an error that causes a disconnect, the payload will go to *both*).

The Error() method will behave as before to decide to reconnect or not.
The AdapterDisconnected() method will check if the disconnect is peer initiated AND a non-error disconnect. If both are true it will invoke the same reconnection checks / actions as Error(). It can't check for all disconnects as ones for an error will go to Error(). Similarly we can't only check in AdapterDisconnected() because all connection failures before a connection is established go to Error, but not AdapterDisconnected()

Refactored the reconnect checks and action to have a single place for both.

Related work items: #3027115