Merged PR 1861781: Updated TestSendMessageAsync for testing sending a binary message only.
This is to have some kind of coverage for sending a binary message. Moved the promise in SendMessage to audio stream session so that we can use it along the way to web_socket layer. Fixed an exception in binary message. Added messageBody in  WebSocketError::UNKNOWN error message so that we can spot what was going on when reading engine log from our clients.

Related work items: #3020805