Merged PR 1875278: fix SIGSEGV in pronunciation assessment
fix SIGSEGV in pronunciation assessment.

The root cause is the service has some random failure where the pronunciation assessment is not included in the result-json, the the `std::stod` will raise SIGSEGV when the string is empty. The fix gives a default value when get the score property.

After the fix, the pronunciation assessment test will still fail when service side doesn't return the score, but will not crash. @<Yinhe Wei> is following the service side fix.

Related work items: #3002584, #3023216