Merged PR 1882488: convert unsupported "other" CGManifest types to Git and Maven entries.
Our CGManifest has a number of components listed as Type: "other".
These entries are not fully scanned, and report a warning in the portal that we do not have full automatic security and legal alerting for these.

I am going to convers as many of the "other" entries into Maven and Git registered entries as I can.  I have one build done, and it pulls in 126 new dependencies, and to my surprise, no new security alerts.

https://msasg.visualstudio.com/Skyman/_componentGovernance/Carbon/history/40136583

Related work items: #3045832