Merged PR 1883299: Add MessageStatus support via ActivityReceived with QS opt-in
## Problem statement
The full background is very long, but the summary: every time a bot does some work (via the processing of an `ITurnContext`), the ConvAI service gets a status depending on what happened--it may have succeeded, it may have failed (exception thrown in the bot), or it may have timed out. These statuses are independent of activities that the bot sends back. We don't currently have any way of receiving these status messages on the client and that's problematic: we can't differentiate between "successful response with *n* activities" and "unsuccessful response after *n* activities," nor can we learn when an operation during a conversation timed out in a non-fatal way.

## Approach
* This is intended to be a fast, minimalistic, and easy-to-modify solution pending a better object model to represent all of this
* A new client config property for opting in to receipt of MessageStatus messages, `DIALOG-RequestBotStatusMessages`, is used to gate all new behavior
* When specified, a QS param, `requestBotStatusMessages=true`, is included on the connection string
* The ConvAI service will use this to decide whether or not these messages are sent
* When received, we statelessly fire the JSON via the string in `ActivityReceivedEventArgs` at which point the explicitly-opted-in client is responsible for checking and processing payload

## Future work
* JS impl -- pending
* Structured impl considerations; e.g. a new `FooBarReceived` event was considered (and I have a functional branch for it)

Related work items: #2702408, #2746758