Merged PR 1899397: verify Doxygen/Swig install fix against master.
Brew deprecated the ability to install a specific version of a package by using a git commit hash.
We use commit hash installs for both SWIG and Doxygen.
Doxygen seems to work with the latest version of the tool, so I switched to just installing latest.
For SWIG, they provide an alternate "Cask/Tap" to install 3.018 from (this is what we need).
I'm switching to that using "brew install swig@3" however, that brings with it the problem that brew no longer links the install in as the default version or puts it in the path.
I first used path hits and passed them into CMAKE, but later found there is a "brew link --force".
This seems to be working and I think it is cleaner that hard coding path hints to pass to each call to CMake .
We should probably see if we can get our build to work with latest version of SWIG, but that is out of scope for this fix.

I should probably spend some time to figure out how to create our own "cask/tap" for Doxygen, because they do not offer official downlevel taps like SWIG does, and Doxygen has broken us in the past, so we should probably have our own TAP that is locked to the version we need, but creating our own TAP seems overly complex (hosting your own fork/mirror of the homebrew-core repo. 😦

I’m still not sure what is broken in “jhakulin/fix-ios-audio-crash”.  I spent a lot of time thinking the failure was because of the version of SWIG we are switching to, but those changes build Mac and iOS fine in master, so there is something else wrong in the audio crash fix branch, and I don’t see what it might be.

Still waiting on nuget tests, but Mac iOS and pythong build tasks are all green.

https://msasg.visualstudio.com/Skyman/_build/results?buildId=16162096&view=results

Related work items: #3061761